+++ src/Game.cpp
@@ -41,7 +41,7 @@ bool Game::is_move_pseudo_legal(const Move& m) const
  * Method: get_pseudo_legal_moves(int square)
  * returns the squares the piece can go to 
  *******************************************************************************/
-std::vector<int> Game::get_pseudo_legal_moves(int square) const
+std::vector<uint8_t> Game::get_pseudo_legal_moves(uint8_t square) const
 {
   return _mgr.get_pseudo_legal_moves(square);
 }
